backend/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
google-cloud-vision==3.4.0
pydantic==1.10.2
python-multipart==0.0.6


absl-py==2.0.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
astunparse==1.6.3
bleach==6.2.0
blinker==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
Faker==33.3.1
fastapi==0.115.6
Flask==3.0.0
flatbuffers==24.12.23
fonttools==4.43.1
gast==0.4.0
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.3
h11==0.14.0
h5py==3.12.1
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kaggle==1.6.17
kagglehub==0.3.6
keras==3.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nest-asyncio==1.5.8
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.14.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==10.0.1
platformdirs==4.0.0
prompt-toolkit==3.0.41
protobuf==5.29.3
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.17.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.9.4
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.13.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.34.0
wcwidth==0.2.10
webencodings==0.5.1
Werkzeug==3.0.0
wrapt==1.14.1
